% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_argsRs.R
\name{get_argsRs}
\alias{get_argsRs}
\title{Title}
\usage{
get_argsRs(dirData, idP = NULL, idU = NULL, equalizeU = FALSE,
  idT = NULL, nTe = 100, pTr = 0.5, minN.P = 2557, dirRefset = NULL,
  classnames = NULL)
}
\arguments{
\item{dirData}{The data directory}

\item{idU}{The ids of the samples from which to sample}

\item{idT}{The ids considered for the test data.}

\item{nTe}{}
}
\description{
Title
}

