% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splitted_refdata_fnctions.R
\name{get_idxUy}
\alias{get_idxUy}
\title{Title}
\usage{
get_idxUy(ids, n, uids = NULL, seed = NULL)
}
\arguments{
\item{ids}{}

\item{n}{}

\item{uids}{}

\item{seed}{}
}
\description{
Title
}

