% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splitted_refdata_fnctions.R
\name{get_idxUy_unlabeled}
\alias{get_idxUy_unlabeled}
\title{Title}
\usage{
get_idxUy_unlabeled(ids, n, uids = NULL, minN = NULL, seed = NULL,
  equalizeU = FALSE)
}
\arguments{
\item{ids}{}

\item{n}{}

\item{uids}{}

\item{minN}{}

\item{seed}{}

\item{equalizeU}{If \code{TRUE}, first a subset 
of the data with equal class priors is created from which  
the unlabeled samples are taken.}
}
\description{
Title
}

