% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parcc.R
\name{get_parcc}
\alias{get_parcc}
\title{Title}
\usage{
get_parcc(seed = NULL, fset = NULL, nP = NULL, nU = NULL, idP = NULL,
  method = NULL, param = NULL, n.seeds = 10,
  dirBase = "D:/Diss/occcomp", dset = "agri6clUforest",
  dirData = paste0(dirBase, "/data/rdata_", dset), dirRes = NULL,
  dirResSum = NULL, nameResSum = NULL, resampling = c("val"),
  scaling = c("ccs01"), get_refset = NULL, args.rs = NULL, nCores = 2)
}
\arguments{
\item{seed}{A vector with the seed points.}

\item{fset}{A vector with the names of the predictor sets. 
(Passed to )}

\item{nP}{}

\item{nU}{}

\item{idP}{The ids considered as positive classes.
unlabeled samples.}

\item{method}{}

\item{param}{}

\item{n.seeds}{}

\item{dirBase}{}

\item{dirData}{}

\item{dirRes}{}

\item{resampling}{}

\item{scaling}{}
}
\description{
Title
}

