% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refset.R
\name{get_refset}
\alias{get_refset}
\title{Get the data for a experiment defined}
\usage{
get_refset(seed, nP, nU, fset, args.rs, overwrite = FALSE, ...)
}
\arguments{
\item{seed}{Seedpoint}

\item{nP}{Number of positive samples}

\item{nU}{Number of unlabeled samples}

\item{fset}{Feature set}

\item{args.rs}{Names list of Arguments required in the functions.}

\item{overwrite}{Overwrite the refset if it is already cached?}

\item{...}{}
}
\value{
data frame with columns set, y, and the feaatures.
}
\description{
Get the data for a experiment defined
}

