% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splitted_refdata_fnctions.R
\name{get_tr_fids_idx}
\alias{get_tr_fids_idx}
\title{Title}
\usage{
get_tr_fids_idx(ids, fids, pTr, seed = NULL)
}
\arguments{
\item{ids}{}

\item{fids}{}

\item{pTr}{}

\item{seed}{}
}
\description{
Title
}

